requirements.txt
cma==3.1.0
gym-microrts==0.5.0
opencv-python>=4.5.5.62
pytorch-lightning==1.5.7
pytorch-metric-learning==1.1.0
seaborn>=0.11.2
stable_baselines3==1.1.0
tensorboard>=2.7.0
tqdm>=4.62.3